requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anyio=4.4.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
brotli-python=1.0.9=py310h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.7.4=py310h06a4308_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
comm=0.2.2=pypi_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.0=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.5.39=0
cuda-runtime=12.4.0=0
cuda-version=12.5=3
cvxopt=1.3.2=pypi_0
debugpy=1.8.2=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
einops=0.8.0=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.20.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.6.1=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
idna=3.7=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pypi_0
ipython=8.26.0=pypi_0
ipywidgets=8.1.3=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=py310h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_2
json5=0.9.25=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.4=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=12.4.2.65=0
libcufft=11.2.0.44=0
libcufile=1.10.1.7=0
libcurand=10.3.6.82=0
libcusolver=11.6.0.99=0
libcusparse=12.3.0.142=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.2.5.2=0
libnvfatbin=12.5.82=0
libnvjitlink=12.4.99=0
libnvjpeg=12.3.1.89=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py310h5eee18b_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mklpy=0.6=pypi_0
monai=1.3.2=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.3=py310h06a4308_0
nibabel=5.2.1=pypi_0
notebook=7.2.1=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_2
openssl=3.0.14=h5eee18b_0
overrides=7.7.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.4.0=py310h5eee18b_0
pip=24.0=py310h06a4308_0
platformdirs=4.2.2=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.47=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pycparser=2.22=pypi_0
pygments=2.18.0=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.14=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=2.0.7=pypi_0
pytorch=2.4.0=py3.10_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_6
pytorch-ignite=0.5.0.post2=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py310h5eee18b_0
pyzmq=26.0.3=pypi_0
qtconsole=5.5.2=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.3=py310h06a4308_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.19.1=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.0=pypi_0
send2trash=1.8.3=pypi_0
setuptools=69.5.1=py310h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
sympy=1.12=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=pypi_0
torchaudio=2.4.0=py310_cu124
torchtriton=3.0.0=py310
torchvision=0.19.0=py310_cu124
tornado=6.4.1=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing_extensions=4.11.0=py310h06a4308_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.2=py310h06a4308_0
vit-pytorch=1.7.4=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.6.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.43.0=py310h06a4308_0
widgetsnbextension=4.0.11=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
